% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{get_bibtex_internal}
\alias{get_bibtex_internal}
\title{Get Citation in bibtext format from DOI}
\usage{
get_bibtex_internal(doi)
}
\description{
This function retrieves the bibtex citation for a given DOI using
the rcrossref package.
}
\keyword{internal}
