% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{get_full_citation_internal}
\alias{get_full_citation_internal}
\title{Get Full Citation in Provided Style from DOI}
\usage{
get_full_citation_internal(doi, style = "apa")
}
\description{
This function retrieves the full citation in provided style for a given DOI using
the rcrossref package.
}
\keyword{internal}
