% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{get_short_citation_internal}
\alias{get_short_citation_internal}
\title{Get citation from DOI}
\usage{
get_short_citation_internal(doi)
}
\arguments{
\item{doi}{The DOI of the publication to cite.}
}
\value{
A formatted citation string.
}
\description{
This function retrieves the citation information for a given DOI.
}
\keyword{internal}
