% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeds.R
\name{get_smarter_breeds}
\alias{get_smarter_breeds}
\title{Get SMARTER Breeds}
\usage{
get_smarter_breeds(query = list(), token = NULL)
}
\arguments{
\item{query}{a \code{list} of query arguments}

\item{token}{a string with a valid token}
}
\value{
Returns a dataframe with selected breeds
}
\description{
Fetch SMARTER REST API breeds endpoint and returns results in a dataframe.
Cached token is used or a new token is generated if not provided when calling
this function (see \code{\link{get_smarter_token}} for more information)
}
\section{Passing additional parameters}{

Breeds endpoint supports additional parameters when making queries.
Additional parameters need to be passed as \code{list} using the \code{query}
parameter. For example, to get all the "Sheep" breeds you need to provide
\code{list(species="Sheep")} as \code{query} parameter. Endpoint supports
query by breed \code{code} and \code{name}. In addition, you can search
using a pattern through the \code{search} parameter.
See \href{https://webserver.ibba.cnr.it/smarter-api/docs/#/Breeds}{
Swagger Breeds endpoint} for more information about the breeds endpoint
}

\examples{
selected_breed <- get_smarter_breeds(query = list(code = "TEX"))

selected_breeds <- get_smarter_breeds(query = list(search = "mer"))
\dontrun{

all_breeds <- get_smarter_breeds()

sheep_breeds <- get_smarter_breeds(query = list(species = "Sheep"))
}
}
