% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{get_smarter_countries}
\alias{get_smarter_countries}
\title{Get SMARTER Countries}
\usage{
get_smarter_countries(query = list())
}
\arguments{
\item{query}{a \code{list} of query arguments}
}
\value{
Returns a dataframe with selected countries
}
\description{
Fetch SMARTER REST API countries endpoint and returns results in a dataframe.
}
\section{Passing additional parameters}{

Contries endpoint supports additional parameters when making queries.
Additional parameters need to be passed as \code{list} using the \code{query}
parameter. For example, to get all the countries where there are "Sheeps"
you need to provide \code{list(species="Sheep")} as \code{query} parameter.
Endpoint supports searches with \emph{two} or \emph{tree} ISO codes using
\code{alpha_2} and \code{alpha3} parameters respectively. In addition,
you can search using a pattern through the \code{search} parameter.
See \href{https://webserver.ibba.cnr.it/smarter-api/docs/#/Countries}{
Swagger Countries endpoint} for more information about the breeds endpoint
}

\examples{
italy <- get_smarter_countries(query = list(name = "Italy"))
france <- get_smarter_countries(query = list(alpha_2 = "FR"))

\dontrun{
# get countries where there are sheeps
sheep_countries <- get_smarter_countries(query = list(species = "Sheep"))
}

}
