% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_smarter_datasets}
\alias{get_smarter_datasets}
\title{Get SMARTER Datasets}
\usage{
get_smarter_datasets(query = list())
}
\arguments{
\item{query}{a \code{list} of query arguments}
}
\value{
Returns a dataframe with selected datasets
}
\description{
Fetch SMARTER REST API datasets endpoint and returns results in a dataframe.
}
\section{Passing additional parameters}{

Datasets endpoint supports additional parameters when making queries.
Additional parameters need to be passed as \code{list} using the \code{query}
parameter. For example, to get all the "Sheep" datasets you need to provide
\code{list(species="Sheep")} as \code{query} parameter. Endpoint supports
query by \code{type} and \code{chip_name}. In addition, you can search
using a pattern in file contents through the \code{search} parameter.
See \href{https://webserver.ibba.cnr.it/smarter-api/docs/#/Datasets}{
Swagger Datasets endpoint} for more information about the datasets endpoint
}

\examples{
genotypes_foreground <- get_smarter_datasets(
  query = list(type = "foreground", type = "genotypes")
)

adaptmap_genotypes <- get_smarter_datasets(
  query = list(species = "Goat", search = "adaptmap", type = "genotypes")
)
\dontrun{

all_datasets <- get_smarter_datasets()
}
}
