% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes.R
\name{get_smarter_genotypes}
\alias{get_smarter_genotypes}
\title{Get SMARTER genotypes}
\usage{
get_smarter_genotypes(species, assembly, dest_path = NULL)
}
\arguments{
\item{species}{a smarter species ("Goat", "Sheep")}

\item{assembly}{the smarter working assembly for such species}

\item{dest_path}{the path where the file will be downloaded. If NULL, the
file will be downloaded in the current working directory}
}
\value{
a character string with the path to the downloaded file is returned
}
\description{
Retrieve genotypes data from SMARTER database. Only information about the
supported assemblies are returned (see \code{\link{get_smarter_info}} for
more information).
}
\section{About SMARTER supported assemblies}{

Currently SMARTER supports \emph{OAR3} and \emph{OAR4} assemblies for Sheep and \emph{CHI1}
and \emph{ARS1} assemblies for goat. Genotypes are provided in PLINK binary
format, and compressed in a zip file.
}

\examples{
\dontrun{
# get genotypes for sheep OAR3 in current directory
get_smarter_genotypes(species = "Sheep", assembly = "OAR3")
}
}
