% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson.R
\name{get_smarter_geojson}
\alias{get_smarter_geojson}
\title{Get SMARTER GeoJSON Samples}
\usage{
get_smarter_geojson(species, token = NULL)
}
\arguments{
\item{species}{a smarter species ("Goat", "Sheep")}

\item{token}{a string with a valid token}
}
\value{
a sf data object
}
\description{
Fetch SMARTER REST API samples with GPS coordinates and return them into as
simple features.
Cached token is used or a new token is generated if not provided when calling
this function (see \code{\link{get_smarter_token}} for more information)
}
\examples{
\dontrun{
# get goat samples with GPS coordinates as sf object
goat_data <- get_smarter_geojson("Goat")

# leaflet doesn't handle MULTIPOINT data (https://github.com/rstudio/leaflet/issues/352)
# Cast them into point (https://r-spatial.github.io/sf/reference/st_cast.html)
goat_data$geometry <- sf::st_cast(goat_data$geometry, "POINT")

# draw samples in a leaflet map using markerCluser
leaflet(data = goat_data) \%>\%
  leaflet::addTiles() \%>\%
  addMarkers(
    clusterOptions = markerClusterOptions(), label = ~smarter_id
  )
}
}
