% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{get_smarter_samples}
\alias{get_smarter_samples}
\title{Get SMARTER Samples}
\usage{
get_smarter_samples(species, query = list())
}
\arguments{
\item{species}{a smarter species ("Goat", "Sheep")}

\item{query}{a \code{list} of query arguments}
}
\value{
Returns a dataframe with selected samples
}
\description{
Fetch SMARTER REST API samples endpoint and returns results in a dataframe.
}
\section{Passing additional parameters}{

Samples endpoint supports additional parameters when making queries.
Additional parameters need to be passed as \code{list} using the \code{query}
parameter. For example, to get all the "foreground" samples you need to
provide \code{list(type="foreground")} as \code{query} parameter.
Endpoint supports a lot of parameters and some of them can be provided
multiple times
See \href{https://webserver.ibba.cnr.it/smarter-api/docs/#/Samples}{
Swagger Samples endpoint} for more information about the samples endpoint
}

\examples{
italian_sheeps <- get_smarter_samples(
  "Sheep",
  query = list(country = "Italy")
)

merino_sheeps <- get_smarter_samples("Sheep", query = list(breed = "Merino"))

selected_goats <- get_smarter_samples(
  "Goat",
  query = list(country = "Italy", breed_code = "ORO", breed_code = "GAR")
)
\dontrun{

foreground_goats <- get_smarter_samples(
  "Goat",
  query = list(type = "foreground")
)

all_sheep_samples <- get_smarter_samples("Sheep")
}
}
