% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_smarter_token}
\alias{get_smarter_token}
\title{Get a SMARTER authentication token}
\usage{
get_smarter_token(
  username = Sys.getenv("SMARTER_API_USERNAME"),
  password = Sys.getenv("SMARTER_API_PASSWORD")
)
}
\arguments{
\item{username}{the SMARTER username (or 'SMARTER_API_USERNAME' environment
variable)}

\item{password}{the SMARTER password (or 'SMARTER_API_PASSWORD' environment
variable)}
}
\value{
Returns a valid token as a string
}
\description{
Authenticate through SMARTER API and returns a valid token. Token is also
cached in package in order to be used by other functions. Calling this
function another time will return the cached token if is still valid
}
\section{Configure environment variables}{

In order to not provide your SMARTER credentials within your code, you could
set those values in your environment. You can call
\code{usethis::edit_r_environ()} in order to track credentials in your
\emph{.Renviron} file using the \code{Key=value} syntax, for example:
\preformatted{
SMARTER_API_USERNAME=<smarter username>
SMARTER_API_PASSWORD=<smarter password>
}
See \href{https://support.rstudio.com/hc/en-us/articles/360047157094-
Managing-R-with-Rprofile-Renviron-Rprofile-site-Renviron-site-rsession-conf-
and-repos-conf}{Managing R with .Rprofile, .Renviron, Rprofile.site,
Renviron.site, rsession.conf, and repos.conf}
for more information
}

\section{Warning}{

SMARTER credentials \strong{MUST NOT be tracked} in your code
}

\examples{
token <- get_smarter_token()
\dontrun{
token <- get_smarter_token(username, password)
}
}
