% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variants.R
\name{get_smarter_variants}
\alias{get_smarter_variants}
\title{Get SMARTER variants}
\usage{
get_smarter_variants(species, assembly, query = list())
}
\arguments{
\item{species}{a smarter species ("Goat", "Sheep")}

\item{assembly}{the smarter working assembly for such species}

\item{query}{a \code{list} of query arguments}
}
\value{
Returns a dataframe with selected variants
}
\description{
Retrieve information on SMARTER SNPs. Only information about the supported
assemblies are returned (see \code{\link{get_smarter_info}} for more
information).
}
\section{About SMARTER supported assemblies}{

Currently SMARTER supports \emph{OAR3} and \emph{OAR4} assemblies for Sheep and \emph{CHI1}
and \emph{ARS1} assemblies for goat. For each assembly, only the evidences used
to create the smarter database are returned. Retrieving all variant for a
certain assembly and species could require time and memory.
}

\examples{
# get variants by region (chrom:start-end)
variants <- get_smarter_variants(
  species = "Goat",
  assembly = "ARS1",
  query = list(region = "1:1-100000")
)
\dontrun{
# get available assemblies from info endpoint
names(get_smarter_info()$working_assemblies)

# collect variants from goat ARS1 endpoint (takes a lot of time)
ars1_variant <- get_smarter_variants(
  species = "Goat",
  assembly = "ARS1"
)
}
}
