% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_chips.R
\name{get_smarter_supportedchips}
\alias{get_smarter_supportedchips}
\title{Get SMARTER supported chips}
\usage{
get_smarter_supportedchips(query = list())
}
\arguments{
\item{query}{a \code{list} of query arguments}
}
\value{
Returns a dataframe with selected chips
}
\description{
Collect information on chip data currently managed in SMARTER database.
}
\examples{
chips <- get_smarter_supportedchips()
\dontrun{
# collect data on chips
get_smarter_supportedchips()

# retrieve samples relying on chip used
samples <- get_smarter_samples(
  species = "Sheep",
  query = list(chip_name = "AffymetrixAxiomOviCan")
)
}
}
